--- app/src/main/java/com/seom/seommain/model/BaseType.kt
@@ -1,4 +1,0 @@
-package com.seom.seommain.model
-
-interface BaseType {
-}
