--- app/src/main/java/com/seom/seommain/model/Model.kt
@@ -1,3 +1,0 @@
-package com.seom.seommain.model
-
-interface Model {}
