+++ app/src/main/java/com/seom/seommain/ui/adapter/mail/MailViewHolder.kt
@@ -1,14 +1,12 @@
-package com.seom.seommain.adapter.mail
+package com.seom.seommain.ui.adapter.mail
 
-import android.util.Log
 import androidx.core.content.ContextCompat
 import androidx.core.view.isGone
 import androidx.core.view.isVisible
 import androidx.recyclerview.widget.RecyclerView
-import com.seom.seommain.R
 import com.seom.seommain.databinding.ItemMailBinding
-import com.seom.seommain.model.mail.MailModel
-import com.seom.seommain.model.mail.ProfileImageType
+import com.seom.seommain.ui.model.mail.MailModel
+import com.seom.seommain.ui.model.mail.ProfileImageType
 
 class MailViewHolder(private val binding: ItemMailBinding) : RecyclerView.ViewHolder(binding.root) {
     fun bind(mail: MailModel) = with(binding) {
