+++ app/src/main/java/com/seom/seommain/ui/model/BaseType.kt
@@ -1,0 +1,4 @@
+package com.seom.seommain.ui.model
+
+interface BaseType {
+}
