+++ app/src/main/java/com/seom/seommain/ui/model/Model.kt
@@ -1,0 +1,3 @@
+package com.seom.seommain.ui.model
+
+interface Model {}
