+++ app/src/main/java/com/seom/seommain/ui/view/MailProfileImage.kt
@@ -1,0 +1,4 @@
+package com.seom.seommain.ui.view
+
+class MailProfileImage {
+}
