+++ app/src/main/java/com/seom/seommain/util/ext/dateExtension.kt
@@ -1,4 +1,4 @@
-package com.seom.seommain.extension
+package com.seom.seommain.util.extension
 
 import android.annotation.SuppressLint
 import java.text.SimpleDateFormat
