--- app/src/main/java/com/seom/seommain/view/MailProfileImage.kt
@@ -1,4 +1,0 @@
-package com.seom.seommain.view
-
-class MailProfileImage {
-}
